+++ tap_zendesk/streams/targets.py
@@ -1,0 +1,11 @@
+from tap_zendesk.streams.abstracts import (
+    PaginatedStream
+)
+
+class Targets(PaginatedStream):
+    name = "targets"
+    replication_method = "FULL_TABLE"
+    key_properties = ["id"]
+    endpoint = 'targets'
+    item_key = 'targets'
+    pagination_type = "offset"