+++ test/unittests/test_request_timeout.py
@@ -1,6 +1,7 @@
 import unittest
-from unittest.mock import AsyncMock, Mock, patch
-from tap_zendesk import http, streams
+from unittest.mock import Mock, patch
+from tap_zendesk import http
+from tap_zendesk.streams import abstracts, TicketAudits
 import requests
 import datetime
 import asyncio
@@ -38,8 +39,7 @@ def mocked_get(*args, **kwargs):
 class TestRequestTimeoutBackoff(unittest.TestCase):
     """
     A set of unit tests to ensure that requests are retrying properly for Timeout Error.
-    """   
-
+    """
     @patch('requests.get')
     def test_call_api_handles_timeout_error(self, mock_get, mock_sleep):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times,
@@ -71,7 +71,7 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         self.assertEqual(mock_get.call_count, 5)
 
     @patch('requests.get', side_effect=[mocked_get(status_code=200, json={"key1": "val1", **PAGINATE_RESPONSE}),
-                                        requests.exceptions.Timeout, requests.exceptions.Timeout, 
+                                        requests.exceptions.Timeout, requests.exceptions.Timeout,
                                         mocked_get(status_code=200, json={"key1": "val1", **SINGLE_RESPONSE})])
     def test_get_cursor_based_handles_timeout_error_in_pagination_call(self, mock_get, mock_sleep):
         """We mock request method to raise a `Timeout`. In next page call the tap should retry request timeout error.
@@ -105,7 +105,7 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         self.assertEqual(mock_get.call_count, 5)
 
     @patch('requests.get', side_effect=[mocked_get(status_code=200, json={"key1": "val1", **PAGINATE_RESPONSE}),
-                                        requests.exceptions.Timeout, requests.exceptions.Timeout, 
+                                        requests.exceptions.Timeout, requests.exceptions.Timeout,
                                         mocked_get(status_code=200, json={"key1": "val1", **SINGLE_RESPONSE})])
     def test_get_offset_based_handles_timeout_error_in_pagination_call(self, mock_get, mock_sleep):
         """We mock request method to raise a `Timeout`. In next page call the tap should retry request timeout error.
@@ -141,10 +141,12 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when `request_timeout` does not passed,
         """
         mock_get.side_effect = requests.exceptions.Timeout
-        cursor_based_stream = streams.CursorBasedStream(config={'subdomain': '34', 'access_token': 'df'})
-        cursor_based_stream.endpoint = 'https://{}'
+
+        paginated_stream = abstracts.PaginatedStream(config={'subdomain': '34', 'access_token': 'df'})
+        paginated_stream.endpoint = 'endpoint_path'
+        paginated_stream.pagination_type = 'cursor'
         try:
-            responses = list(cursor_based_stream.get_objects())
+            responses = list(paginated_stream.get_objects())
         except requests.exceptions.Timeout as e:
             pass
 
@@ -156,10 +158,11 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when string "0" value of `request_timeout` passed,
         """
         mock_get.side_effect = requests.exceptions.Timeout
-        cursor_based_stream = streams.CursorBasedStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': "0"})
-        cursor_based_stream.endpoint = 'https://{}'
+        paginated_stream = abstracts.PaginatedStream(config={'subdomain': '34', 'access_token': 'df'})
+        paginated_stream.endpoint = 'endpoint_path'
+        paginated_stream.pagination_type = 'cursor'
         try:
-            responses = list(cursor_based_stream.get_objects())
+            responses = list(paginated_stream.get_objects())
         except requests.exceptions.Timeout as e:
             pass
 
@@ -171,10 +174,11 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when int 0 value of `request_timeout` passed,
         """
         mock_get.side_effect = requests.exceptions.Timeout
-        cursor_based_stream = streams.CursorBasedStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': 0})
-        cursor_based_stream.endpoint = 'https://{}'
+        paginated_stream = abstracts.PaginatedStream(config={'subdomain': '34', 'access_token': 'df'})
+        paginated_stream.endpoint = 'endpoint_path'
+        paginated_stream.pagination_type = 'cursor'
         try:
-            responses = list(cursor_based_stream.get_objects())
+            responses = list(paginated_stream.get_objects())
         except requests.exceptions.Timeout as e:
             pass
 
@@ -186,10 +190,11 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when string value of `request_timeout` passed,
         """
         mock_get.side_effect = requests.exceptions.Timeout
-        cursor_based_stream = streams.CursorBasedStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': REQUEST_TIMEOUT_STR})
-        cursor_based_stream.endpoint = 'https://{}'
+        paginated_stream = abstracts.PaginatedStream(config={'subdomain': '34', 'access_token': 'df'})
+        paginated_stream.endpoint = 'endpoint_path'
+        paginated_stream.pagination_type = 'cursor'
         try:
-            responses = list(cursor_based_stream.get_objects())
+            responses = list(paginated_stream.get_objects())
         except requests.exceptions.Timeout as e:
             pass
 
@@ -201,10 +206,11 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when int value of `request_timeout` passed,
         """
         mock_get.side_effect = requests.exceptions.Timeout
-        cursor_based_stream = streams.CursorBasedStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': REQUEST_TIMEOUT})
-        cursor_based_stream.endpoint = 'https://{}'
+        paginated_stream = abstracts.PaginatedStream(config={'subdomain': '34', 'access_token': 'df'})
+        paginated_stream.endpoint = 'endpoint_path'
+        paginated_stream.pagination_type = 'cursor'
         try:
-            responses = list(cursor_based_stream.get_objects())
+            responses = list(paginated_stream.get_objects())
         except requests.exceptions.Timeout as e:
             pass
 
@@ -216,10 +222,11 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when float value of `request_timeout` passed,
         """
         mock_get.side_effect = requests.exceptions.Timeout
-        cursor_based_stream = streams.CursorBasedStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': REQUEST_TIMEOUT_FLOAT})
-        cursor_based_stream.endpoint = 'https://{}'
+        paginated_stream = abstracts.PaginatedStream(config={'subdomain': '34', 'access_token': 'df'})
+        paginated_stream.endpoint = 'endpoint_path'
+        paginated_stream.pagination_type = 'cursor'
         try:
-            responses = list(cursor_based_stream.get_objects())
+            responses = list(paginated_stream.get_objects())
         except requests.exceptions.Timeout as e:
             pass
 
@@ -230,10 +237,11 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when empty value of `request_timeout` passed,
         """
         mock_get.side_effect = requests.exceptions.Timeout
-        cursor_based_stream = streams.CursorBasedStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': ''})
-        cursor_based_stream.endpoint = 'https://{}'
+        paginated_stream = abstracts.PaginatedStream(config={'subdomain': '34', 'access_token': 'df'})
+        paginated_stream.endpoint = 'endpoint_path'
+        paginated_stream.pagination_type = 'cursor'
         try:
-            responses = list(cursor_based_stream.get_objects())
+            responses = list(paginated_stream.get_objects())
         except requests.exceptions.Timeout as e:
             pass
 
@@ -244,8 +252,8 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when `request_timeout` does not passed,
         """
         mock_get.side_effect = requests.exceptions.Timeout
-        cursor_based_export_stream = streams.CursorBasedExportStream(config={'subdomain': '34', 'access_token': 'df'})
-        cursor_based_export_stream.endpoint = 'https://{}'
+        cursor_based_export_stream = abstracts.CursorBasedExportStream(config={'subdomain': '34', 'access_token': 'df'})
+        cursor_based_export_stream.endpoint = 'endpoint_path'
         try:
             responses = list(cursor_based_export_stream.get_objects(datetime.datetime.utcnow()))
         except requests.exceptions.Timeout as e:
@@ -259,8 +267,8 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when stiring "0" value of `request_timeout` passed,
         """
         mock_get.side_effect = requests.exceptions.Timeout
-        cursor_based_export_stream = streams.CursorBasedExportStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': '0'})
-        cursor_based_export_stream.endpoint = 'https://{}'
+        cursor_based_export_stream = abstracts.CursorBasedExportStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': '0'})
+        cursor_based_export_stream.endpoint = 'endpoint_path'
         try:
             responses = list(cursor_based_export_stream.get_objects(datetime.datetime.utcnow()))
         except requests.exceptions.Timeout as e:
@@ -274,8 +282,8 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when int 0 value of `request_timeout` passed,
         """
         mock_get.side_effect = requests.exceptions.Timeout
-        cursor_based_export_stream = streams.CursorBasedExportStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': 0})
-        cursor_based_export_stream.endpoint = 'https://{}'
+        cursor_based_export_stream = abstracts.CursorBasedExportStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': 0})
+        cursor_based_export_stream.endpoint = 'endpoint_path'
         try:
             responses = list(cursor_based_export_stream.get_objects(datetime.datetime.utcnow()))
         except requests.exceptions.Timeout as e:
@@ -289,8 +297,8 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when empty value of `request_timeout` passed,
         """
         mock_get.side_effect = requests.exceptions.Timeout
-        cursor_based_export_stream = streams.CursorBasedExportStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': ''})
-        cursor_based_export_stream.endpoint = 'https://{}'
+        cursor_based_export_stream = abstracts.CursorBasedExportStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': ''})
+        cursor_based_export_stream.endpoint = 'endpoint_path'
         try:
             responses = list(cursor_based_export_stream.get_objects(datetime.datetime.utcnow()))
         except requests.exceptions.Timeout as e:
@@ -304,8 +312,8 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when string value of `request_timeout` passed,
         """
         mock_get.side_effect = requests.exceptions.Timeout
-        cursor_based_export_stream = streams.CursorBasedExportStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': REQUEST_TIMEOUT_STR})
-        cursor_based_export_stream.endpoint = 'https://{}'
+        cursor_based_export_stream = abstracts.CursorBasedExportStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': REQUEST_TIMEOUT_STR})
+        cursor_based_export_stream.endpoint = 'endpoint_path'
         try:
             responses = list(cursor_based_export_stream.get_objects(datetime.datetime.utcnow()))
         except requests.exceptions.Timeout as e:
@@ -319,8 +327,8 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when int value of `request_timeout` passed,
         """
         mock_get.side_effect = requests.exceptions.Timeout
-        cursor_based_export_stream = streams.CursorBasedExportStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': REQUEST_TIMEOUT})
-        cursor_based_export_stream.endpoint = 'https://{}'
+        cursor_based_export_stream = abstracts.CursorBasedExportStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': REQUEST_TIMEOUT})
+        cursor_based_export_stream.endpoint = 'endpoint_path'
         try:
             responses = list(cursor_based_export_stream.get_objects(datetime.datetime.utcnow()))
         except requests.exceptions.Timeout as e:
@@ -334,8 +342,8 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when float value of `request_timeout` passed,
         """
         mock_get.side_effect = requests.exceptions.Timeout
-        cursor_based_export_stream = streams.CursorBasedExportStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': REQUEST_TIMEOUT_FLOAT})
-        cursor_based_export_stream.endpoint = 'https://{}'
+        cursor_based_export_stream = abstracts.CursorBasedExportStream(config={'subdomain': '34', 'access_token': 'df', 'request_timeout': REQUEST_TIMEOUT_FLOAT})
+        cursor_based_export_stream.endpoint = 'endpoint_path'
         try:
             responses = list(cursor_based_export_stream.get_objects(datetime.datetime.utcnow()))
         except requests.exceptions.Timeout as e:
@@ -352,7 +360,7 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when `request_timeout` does not passed,"""
         mock_get.return_value.__aenter__.side_effect = requests.exceptions.Timeout
 
-        ticket_audits = streams.TicketAudits(
+        ticket_audits = TicketAudits(
             config={"subdomain": "test-zendesk", "access_token": "df"}
         )
 
@@ -375,7 +383,7 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
     ):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when string "0" value of `request_timeout` passed,"""
         mock_get.return_value.__aenter__.side_effect = requests.exceptions.Timeout
-        ticket_audits = streams.TicketAudits(
+        ticket_audits = TicketAudits(
             config={
                 "subdomain": "test-zendesk",
                 "access_token": "df",
@@ -403,7 +411,7 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when int 0 value of `request_timeout` passed,"""
         mock_get.return_value.__aenter__.side_effect = requests.exceptions.Timeout
 
-        ticket_audits = streams.TicketAudits(
+        ticket_audits = TicketAudits(
             config={
                 "subdomain": "test-zendesk",
                 "access_token": "df",
@@ -431,7 +439,7 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when string value of `request_timeout` passed,"""
         mock_get.return_value.__aenter__.side_effect = requests.exceptions.Timeout
 
-        ticket_audits = streams.TicketAudits(
+        ticket_audits = TicketAudits(
             config={
                 "subdomain": "test-zendesk",
                 "access_token": "df",
@@ -459,7 +467,7 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when int value of `request_timeout` passed,"""
         mock_get.return_value.__aenter__.side_effect = requests.exceptions.Timeout
 
-        ticket_audits = streams.TicketAudits(
+        ticket_audits = TicketAudits(
             config={
                 "subdomain": "test-zendesk",
                 "access_token": "df",
@@ -487,7 +495,7 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when float value of `request_timeout` passed,"""
         mock_get.return_value.__aenter__.side_effect = requests.exceptions.Timeout
 
-        ticket_audits = streams.TicketAudits(
+        ticket_audits = TicketAudits(
             config={
                 "subdomain": "test-zendesk",
                 "access_token": "df",
@@ -515,7 +523,7 @@ class TestRequestTimeoutBackoff(unittest.TestCase):
         """We mock request method to raise a `Timeout` and expect the tap to retry this up to 5 times when empty value of `request_timeout` passed,"""
         mock_get.return_value.__aenter__.side_effect = requests.exceptions.Timeout
 
-        ticket_audits = streams.TicketAudits(
+        ticket_audits = TicketAudits(
             config={
                 "subdomain": "test-zendesk",
                 "access_token": "df",