+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.9.0
+  * Added multiple new streams [#176](https://github.com/singer-io/tap-zendesk/pull/176)
+  * bump singer-python to 6.8.0, zenpy to 2.0.57 and requests to 2.34.2
+
 # 2.8.6
   * Schema updates [#199](https://github.com/singer-io/tap-zendesk/pull/199)